logger/requirements.txt
# Python Logger Dependencies
# Install: pip install -r requirements.txt

# MQTT client
paho-mqtt>=2.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization (optional, for histogram)
matplotlib>=3.7.0

# Environment variables (optional)
python-dotenv>=1.0.0
